% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{caller_name.str}
\alias{caller_name.str}
\title{Finding out who called the parent function}
\usage{
caller_name.str(n = 2)
}
\value{
caller of the parent function (i.e call of caller)
}
\description{
Finding out who called the parent function
}
