% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{grab_grob}
\alias{grab_grob}
\title{Grabs the last drawing as a Grob and returns it (to be used by other Grob things)}
\usage{
grab_grob()
}
\value{
a grob
}
\description{
Grabs the last drawing as a Grob and returns it (to be used by other Grob things)
}
