% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_room.R
\name{reportage}
\alias{reportage}
\title{Applies a function on a dataframe, with specific response columns, and report on the specified variables from the output}
\usage{
reportage(data.df, all.response.names.v, the.method, report.items.v, ...)
}
\arguments{
\item{report.items.v}{}
}
\description{
Applies a function on a dataframe, with specific response columns, and report on the specified variables from the output
}
