% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{sv_as_factor.dataframe}
\alias{sv_as_factor.dataframe}
\title{Converts a data.frame to a factor data.frame each var with the.levels level}
\usage{
sv_as_factor.dataframe(data.df, the.levels, be.ordered = TRUE)
}
\arguments{
\item{be.ordered}{}
}
\description{
Converts a data.frame to a factor data.frame each var with the.levels level
}
