% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{sv_as_integer.dataframe}
\alias{sv_as_integer.dataframe}
\title{Converts a data.frame to an integer data.frame}
\usage{
sv_as_integer.dataframe(data.df)
}
\arguments{
\item{data.df}{}
}
\description{
Converts a data.frame to an integer data.frame
}
