% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_factor.R
\name{sv_factor_mix}
\alias{sv_factor_mix}
\title{average of the factors}
\usage{
sv_factor_mix(the.levels, ...)
}
\arguments{
\item{the.levels}{The number of levels in the output.
It cannot be less than the maximum number in the input factos.
A value of -1 means it should be adjusted to the maximum number in the input.}

\item{...}{}
}
\description{
average of the factors
}
