% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lasso.R
\name{sv_lasso}
\alias{sv_lasso}
\title{Do all LASSO things}
\usage{
sv_lasso(data.df, response.name, the.response.mode = c("gaussian",
  "mgaussian", "ordinal", "bionomial", "multinomial"), best.lambda = NULL,
  alpha = 1, nlambda = 100, type.measure = "deviance", file.prefix = "",
  ...)
}
\description{
Do all LASSO things
}
