% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_unification.R
\name{sv_meta_compare}
\alias{sv_meta_compare}
\title{compares two data.frames returns the variables are different between them}
\usage{
sv_meta_compare(the.data.1, the.data.2)
}
\arguments{
\item{the.data.2}{}
}
\description{
compares two data.frames returns the variables are different between them
}
