% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_unification.R
\name{sv_meta_harmonize}
\alias{sv_meta_harmonize}
\title{Title}
\usage{
sv_meta_harmonize(the.data, the.type = c("label", "value.labels",
  "value.values"), the.mains.and.alternatives.ls)
}
\arguments{
\item{the.mains.and.alternatives.ls}{}
}
\description{
Title
}
