% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_spss.R
\name{sv_meta_variables.df}
\alias{sv_meta_variables.df}
\title{Reports each variable, its possible values, the labels of those possible values}
\usage{
sv_meta_variables.df(the.data)
}
\arguments{
\item{the.data}{}
}
\description{
Reports each variable, its possible values, the labels of those possible values
}
