% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_spss.R
\name{sv_meta_variables_detailed}
\alias{sv_meta_variables_detailed}
\title{Returns structural reports on the dataset}
\usage{
sv_meta_variables_detailed(data.df, the.origin = c("spss", "r"))
}
\arguments{
\item{data.df}{}
}
\description{
Returns structural reports on the dataset
}
