% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{sv_plot_corr_mat}
\alias{sv_plot_corr_mat}
\title{Plot correlation matrix}
\usage{
sv_plot_corr_mat(the.data, calc.method = c("kendall", "spearman", "pearson"),
  conf.level = 0.95, miss.method = c("pairwise.complete.obs", "everything",
  "complete.obs", "all.obs", "na.or.complete"), group_by = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot correlation matrix
}
