% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_plssem.R
\name{sv_plssem}
\alias{sv_plssem}
\title{Do PLS-SEM}
\usage{
sv_plssem(data.df, struct.mat, measure.mat, the.model, the.lv,
  the.filename = "plssem_result", show.rsquared = TRUE)
}
\arguments{
\item{the.model}{If model is presented then there would be a MatrixPLS result}

\item{show.rsquared}{}
}
\description{
Do PLS-SEM
}
