% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_clusters.R
\name{sv_report_cluster}
\alias{sv_report_cluster}
\title{Reports on the cluster}
\usage{
sv_report_cluster(data.df, the.name, to.impute = FALSE)
}
\arguments{
\item{data.df}{The input data}

\item{the.name}{The name of the report (file)}
}
\description{
Reports on the cluster
}
