% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_room.R
\name{sv_reportage}
\alias{sv_reportage}
\title{Applies a function on a dataframe, with specific response columns, and reports on the specified variables from the output}
\usage{
sv_reportage(data.df, all.response.names.v, the.method, report.items.v,
  response.selection = c("all.out.except.one", "all.in.except.one"),
  draw.plots = FALSE, plots.prefix = NULL, ...)
}
\arguments{
\item{data.df}{data input}

\item{all.response.names.v}{the variables in the data inputh which are the responses}

\item{the.method}{the method to be called}

\item{report.items.v}{what items to be returned back from the the.method result (specific to the.method)}

\item{response.selection}{extra variables to be pass to the the.method}

\item{draw.plots}{will draw plots}

\item{plots.prefix}{what name attached at the begining of plot files}
}
\description{
Applies a function on a dataframe, with specific response columns, and reports on the specified variables from the output
}
