% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{sv_set_scenario}
\alias{sv_set_scenario}
\title{Keep track of the active scenarios name and give a clojure to check them later}
\usage{
sv_set_scenario(...)
}
\arguments{
\item{...}{a list of scenarios (probably their names)}
}
\value{
a clojure function that gets a list of scenarios (probably their names) and return TRUE/FALSE if they are in the set
}
\description{
Keep track of the active scenarios name and give a clojure to check them later
}
\examples{
is_there <- sv_set_scenario('a', 'b')
print(is_there('a')) # prints TRUE
}
