% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{uv_list_to_df}
\alias{uv_list_to_df}
\title{Converts a list to a dataframe with 'list.element' and 'names' columns}
\usage{
uv_list_to_df(listfordf)
}
\arguments{
\item{listfordf}{}
}
\description{
Converts a list to a dataframe with 'list.element' and 'names' columns
}
