% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_unification.R
\name{uv_meta_change_value_label}
\alias{uv_meta_change_value_label}
\title{Title}
\usage{
uv_meta_change_value_label(the.data, the.variable, the.value.label.src,
  the.value.label.des)
}
\arguments{
\item{the.value.label.des}{}
}
\description{
Title
}
