% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_structure.R
\name{uv_meta_spss_variables.df}
\alias{uv_meta_spss_variables.df}
\title{Reports each variable, its possible values, the labels of those possible values}
\usage{
uv_meta_spss_variables.df(the.data)
}
\arguments{
\item{the.data}{}
}
\description{
Reports each variable, its possible values, the labels of those possible values
}
