% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_preparation.R
\name{uv_test_and_train_sets}
\alias{uv_test_and_train_sets}
\title{prepares 4 datasets : train and test both in predictors and response}
\usage{
uv_test_and_train_sets(data.df, response.name, the.train.ratio = 0.75,
  pred.vars.handling = c("number", "dummify"),
  resp.vars.handling = c("2num"))
}
\arguments{
\item{data.df}{input data}

\item{response.name}{the name of the column which is the response}

\item{the.train.ratio}{the percent of train data to the whole data (the rest for test)}

\item{pred.vars.handling}{how to handle predictor variables. It can be 'number' or 'dummify'}

\item{resp.vars.handling}{how to handle response variables. It can be 'number' or 'dummify'}
}
\value{
Returns a list of 4 dataset for train and test both in predictors and response
}
\description{
prepares 4 datasets : train and test both in predictors and response
}
