% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{wrap_plot_call}
\alias{wrap_plot_call}
\title{Wraps a plotting function with required device settings}
\usage{
wrap_plot_call(plot_fun, the.name = "")
}
\description{
Wraps a plotting function with required device settings
}
